% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples_sgflmm.R
\name{posterior_samples_sgflmm}
\alias{posterior_samples_sgflmm}
\title{Bayesian inference for spatial generalized functional linear mixed model (SGFLMM)}
\usage{
posterior_samples_sgflmm(
  M,
  initial,
  proposal_var_fixed,
  proposal_var_sigma2,
  proposal_var_rand,
  autotune,
  prior_a,
  prior_b,
  M_auxz,
  data_list,
  nbd_list,
  X_list,
  t,
  p,
  basis_type = "fourier"
)
}
\arguments{
\item{M}{An integer specifying the number of MCMC iterations to perform.}

\item{initial}{A (p+G+3)-dimsional vector giving the initial values for the parameters 
(\eqn{\eta}, \eqn{\alpha}, trunctated coefficients \eqn{\beta_j=\int \beta(t) \phi_j(t) dt}, j=1,...,p,
\eqn{u_g}, g= 1,..., G, \eqn{\sigma_u^2}).}

\item{proposal_var_fixed}{A positive numeric value specifying the diagonal proposal 
variance used for the fixed effects (\eqn{\alpha} and \eqn{\beta_j}, j=1,...,p) in the random walk 
Metropolis-Hastings algorithm.}

\item{proposal_var_sigma2}{A positive numeric value specifying the diagonal proposal 
variance used for \eqn{\sigma_u^2} in the random walk Metropolis-Hastings algorithm.}

\item{proposal_var_rand}{A positive numeric value specifying the diagonal proposal 
variance used for the random effects (\eqn{u_g}, g=1,...,G) in the random walk 
Metropolis-Hastings algorithm.}

\item{autotune}{Logical; if \code{TRUE}, the sampler adaptively adjusts 
\code{proposal_var_fixed}, \code{proposal_var_sigma2}, and \code{proposal_var_rand} 
based on the acceptance rate during sampling. The 
variance is increased if the acceptance rate is high (above 35\%) and decreased 
if the rate is too low (below 20\%). If \code{FALSE}, the proposal variance 
remains fixed throughout the sampling process.}

\item{prior_a}{Shape parameter of the inverse-gamma prior for \eqn{\sigma_u^2}.}

\item{prior_b}{Scale parameter of the inverse-gamma prior for \eqn{\sigma_u^2}.}

\item{M_auxz}{The number of full Gibbs sampling used to generate the auxiliary variable.}

\item{data_list}{A list of length G, where each element contains the binary response variable for a corresponding group.}

\item{nbd_list}{A list of length G, where each element contains the neighborhood structure (of class \code{nb}) for a corresponding group.}

\item{X_list}{A list of length G, where each element is an \eqn{n_g \times T} matrix of functional covariates for group \eqn{g}. 
Each row corresponds to one observed functional covariate.}

\item{t}{A T-dimensional vector containing densely time grid points on the interval \eqn{[0, 1]}, 
where the functions are evaluated.
This should match the time domain used when creating basis functions via the \code{CreateBasis} function 
from the \pkg{fdapace} package.}

\item{p}{An natural number indicating the truncation level.}

\item{basis_type}{A character string specifying the type of orthogonal basis \eqn{\phi_j, j=1,..., p}. 
Must be one of \code{"cos"}, \code{"sin"}, \code{"fourier"}, \code{"legendre01"}, or \code{"poly"}. 
These correspond to basis types supported by the \code{CreateBasis} function in the \pkg{fdapace} package.
The default is \code{"fourier"}.}
}
\value{
A list with the following elements:
\describe{
 \item{samples}{A list containing posterior samples:
   \itemize{
      \item \code{eta}: An M by 1 matrix of posterior samples for the spatial dependence parameter.
      \item \code{alpha}: An M by 1 matrix of posterior samples for the intercept parameter.
      \item \code{beta}: An M by p matrix of posterior samples for the trunctated coefficients \eqn{\beta_j}, j=1,...,p.
      \item \code{rand}: An M by G matrix of posterior samples for the random effect.
      \item \code{sigma2}: An M by 1 matrix of posterior samples for the variance component of the random effect.
   }
 }
  \item{diagnostics}{A list containing diagnostic information:
    \itemize{
     \item \code{final_proposal_var_fixed}: The final proposal variance used for the fixed effect.
      \item \code{accept_rate_fixed}: The acceptance rate for the fixed effect proposals. 
      \item \code{final_proposal_var_rand}: The final proposal variance used for the random effect.
      \item \code{accept_rate_rand}: The acceptance rate for the random effect proposals. 
      \item \code{final_proposal_var_sigma2}: The final proposal variance used for the variance component of the random effect.
      \item \code{accept_rate_sigma2}: The acceptance rate for the variance component of the random effect.
      \item \code{elapsed_time_sec}: The total elapsed computation time in seconds.
    }
  }
}
}
\description{
Performs Bayesian inference for a spatial generalized functional linear mixed model (SGFLMM), 
using Metropolis-Hastings sampling with a random walk proposal distribution. 
The proposal variance can be fixed or adaptively tuned
during sampling to achieve efficient exploration of the posterior distribution.
}
\examples{

library(SGFLMMBayesian)

# Load example data included in the package
data(data_example)
X_list <- data_example$X_list
data_list <- data_example$data_list
nbd_list <- data_example$nbd_list

# simulate time point on [0, 1]
t <- seq(from = 0, to = 1, length.out = ncol(X_list[[1]]))

initial_values <- c(0.5, -4.6, 1.1, -0.9, 2.2, -1.4, 0.4, 0.0, 0.1, -2.5, 2.0, -0.2, -1.1, -0.3, 1.0, 0.2, 0.4, 1.4)
res <- posterior_samples_sgflmm(M=8000, initial=initial_values, 
proposal_var_fixed=0.001, proposal_var_sigma2=1.1, proposal_var_rand=1, autotune=FALSE, 
prior_a=1, prior_b=1, M_auxz = 20, 
data_list=data_list, nbd_list = nbd_list,
X_list = X_list, t=t, p=3, basis_type="fourier") 

# Example: compute the posterior mean of eta
mean(res$samples$eta[-(1:4000)])

}
\seealso{
\code{\link{posterior_samples_gflm}}
\code{\link{posterior_samples_sgflm}}
\code{\link{posterior_samples_gflmm}}
}
