\name{SGFLMMBayesian-package}
\alias{SGFLMMBayesian-package}
\alias{SGFLMMBayesian}
\docType{package}
\title{
\packageTitle{SGFLMMBayesian}
}
\description{
\packageDescription{SGFLMMBayesian}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{SGFLMMBayesian}

Maintainer: \packageMaintainer{SGFLMMBayesian}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = SGFLMMBayesian)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{SGFLMMBayesian}}
%\section{Documentation Index}{\packageIndices{SGFLMMBayesian}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
